removed dogTasks files and added dogProfile files
