+++ controllers/dogProfilesController.js
@@ -1,1 +1,36 @@
 const db = require("../models");
+
+// Defining methods for the dogProfilesController
+module.exports = {
+    findAll: function(req, res) {
+      db.dogProfile
+        .find(req.query)
+        .then(dbModel => res.json(dbModel))
+        .catch(err => res.status(422).json(err));
+    },
+    findById: function(req, res) {
+      db.dogProfile
+        .findById(req.params.id)
+        .then(dbModel => res.json(dbModel))
+        .catch(err => res.status(422).json(err));
+    },
+    create: function(req, res) {
+      db.dogProfile
+        .create(req.body)
+        .then(dbModel => res.json(dbModel))
+        .catch(err => res.status(422).json(err));
+    },
+    update: function(req, res) {
+      db.dogProfile
+        .findOneAndUpdate({ _id: req.params.id }, req.body)
+        .then(dbModel => res.json(dbModel))
+        .catch(err => res.status(422).json(err));
+    },
+    remove: function(req, res) {
+      db.dogProfile
+        .findById({ _id: req.params.id })
+        .then(dbModel => dbModel.remove())
+        .then(dbModel => res.json(dbModel))
+        .catch(err => res.status(422).json(err));
+    }
+  };
