--- controllers/dogTasksController.js
@@ -1,36 +1,0 @@
-const db = require("../models");
-
-// Defining methods for the dogTasksController
-module.exports = {
-    findAll: function(req, res) {
-      db.dogTask
-        .find(req.query)
-        .then(dbModel => res.json(dbModel))
-        .catch(err => res.status(422).json(err));
-    },
-    findById: function(req, res) {
-      db.dogTask
-        .findById(req.params.id)
-        .then(dbModel => res.json(dbModel))
-        .catch(err => res.status(422).json(err));
-    },
-    create: function(req, res) {
-      db.dogTask
-        .create(req.body)
-        .then(dbModel => res.json(dbModel))
-        .catch(err => res.status(422).json(err));
-    },
-    update: function(req, res) {
-      db.dogTask
-        .findOneAndUpdate({ _id: req.params.id }, req.body)
-        .then(dbModel => res.json(dbModel))
-        .catch(err => res.status(422).json(err));
-    },
-    remove: function(req, res) {
-      db.dogTask
-        .findById({ _id: req.params.id })
-        .then(dbModel => dbModel.remove())
-        .then(dbModel => res.json(dbModel))
-        .catch(err => res.status(422).json(err));
-    }
-  };
